git-svn-id: http://10.205.116.236/svn/AIR@193 d2170373-ae6f-44b7-9150-2eec3990abfc

+++ branches/1.10/src/com/bayerbbs/applrepos/hibernate/HistoryHbn.java
@@ -182,7 +182,7 @@ public class HistoryHbn {
 		Query selectQuery = session.createSQLQuery("SELECT SEQ_HISTORY_EVENT.NEXTVAL FROM DUAL");
 		try 
 		{
-			result = (Long) ((BigDecimal) selectQuery.uniqueResult()).longValue();
+			result = ((BigDecimal) selectQuery.uniqueResult()).longValue();
 		} 
 		catch (Exception e) 
 		{
